+++ src/functions/set/difference.js
@@ -3,12 +3,12 @@ const LimeFunction = require('../../structs/function');
 
 /* ------------------------ division ------------------------ */
 
-// Intersection function class
-class LimeFunctionIntersection extends LimeFunction {
+// Difference function class
+class LimeFunctionDifference extends LimeFunction {
   // Constructor
   constructor(lime, mode) {
     // Super from function class
-    super(lime, { name: 'intersection', mode });
+    super(lime, { name: 'difference', mode });
 
     // Operations
     this.operations.r = [
@@ -18,7 +18,7 @@ class LimeFunctionIntersection extends LimeFunction {
     // Algorithms
     this.algorithms.set('r(arg{set[2]})', (step) => {
       const l = step.right.places[0]; const r = step.right.places[1];
-      const places = [...l.places].filter((u) => r.places.some((v) => this.lime.direct([u, '==', v]).value));
+      const places = [...l.places].filter((u) => r.places.every((v) => this.lime.direct([u, '!=', v]).value));
       step.rus(this.lime.build('set')(...places));
     });
   }
@@ -27,4 +27,4 @@ class LimeFunctionIntersection extends LimeFunction {
 /* ------------------------ division ------------------------ */
 
 // Export
-module.exports = LimeFunctionIntersection;
+module.exports = LimeFunctionDifference;