+++ src/functions/set/union.js
@@ -17,14 +17,9 @@ class LimeFunctionUnion extends LimeFunction {
 
     // Algorithms
     this.algorithms.set('r(arg{set[2]})', (step) => {
-      const places = step.right.places.map((s) => s.places).flat();
-      const t = this.lime.direct(['intersection', '(', step.right.places[0], ',', step.right.places[1], ')']).places;
-      for (let i = 0; i < places.length; i++) {
-        if (t.some((v) => this.lime.direct([places[i], '==', v]).value)) {
-          places.splice(i, 1);
-        }
-      }
-      step.rus(this.lime.build('set')(...places, ...t));
+      const s = this.lime.direct(['symdiff', '(', step.right.places[0], ',', step.right.places[1], ')']);
+      const i = this.lime.direct(['intersection', '(', step.right.places[0], ',', step.right.places[1], ')']);
+      step.rus(this.lime.build('set')(...s.places, ...i.places));
     });
   }
 }